+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ## [Unreleased]
 ### Added
 - Check that test classes do not define mutable fields.
+- Check that test classes do not define mutable properties
 ### Fixed
 ### Changed
 ### Removed